Fix compilation error on old compilers
git-svn-id: https://subversion.gsi.de/dabc/trunk@2632 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ plugins/fesa/src/Player.cxx
@@ -229,7 +229,7 @@ void fesa::Player::ProcessTimerEvent(unsigned timer)
 
 #ifdef WITH_ROOT
 
-   fWorkerHierarchy.GetHChild("StreamerInfo").SetField(dabc::prop_hash, fSniffer->GetStreamerInfoHash());
+   fWorkerHierarchy.GetHChild("StreamerInfo").SetField(dabc::prop_hash, (uint64_t) fSniffer->GetStreamerInfoHash());
 
    TH2I* h2 = (TH2I*) fHist;
    if (h2!=0) {
